[Resolves #568] update-stack and update-stack-cs return code
The update-stack and update-stack-cs command returns a non 0 status
when there are no changes to peform.  In a CI/CD scenario we want
those commands to exit with 0 return status in this scenario.

Note - CF will put stack sets in a 'defuct' state when there is
not changes to perform.